% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculateFWERforRank}
\alias{calculateFWERforRank}
\title{Calculate Family-Wise Error Rate values for each rank}
\usage{
calculateFWERforRank(truevalues, permutedvalues, quantiledata, alpha = 0.05)
}
\arguments{
\item{truevalues}{a tibble of featureRanks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_true_fws function}

\item{permutedvalues}{a tibble of featureRanks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}

\item{alpha}{The alpha value for p-value calculation, default = 0.05}
}
\value{
A dataframe containing FWER for each permutation
}
\description{
Calculate Family-Wise Error Rate values for each rank
}
