% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculatePvalueforRank}
\alias{calculatePvalueforRank}
\title{Calculate P value for each rank}
\usage{
calculatePvalueforRank(truevalues, permutatedvalues, quantiledata)
}
\arguments{
\item{truevalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_true_fws function}

\item{permutatedvalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_permutated_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A dataframe of featureRanks, proportions and pvalues
}
\description{
Calculate P value for each rank
}
