% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculatePvalueforSet}
\alias{calculatePvalueforSet}
\title{Calculate P value for entire feature set}
\usage{
calculatePvalueforSet(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{a tibble of featureRanks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A dataframe with one column and one row in which contains the proportion-based p-value for the entire feature set.
}
\description{
Calculate P value for entire feature set
}
