% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalues.R
\name{calculateQuantiles}
\alias{calculateQuantiles}
\title{Calculate Quantiles}
\usage{
calculateQuantiles(truevalues, permutedvalues)
}
\arguments{
\item{truevalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_true_fws function}

\item{permutedvalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_permuted_fws function}
}
\value{
A tibble of quantile data.
}
\description{
Calculate Quantiles
}
