% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{fiRankPlot}
\alias{fiRankPlot}
\title{Generate Feature Importance Score-Rank Plot
This function generates a figure which shows the true observed data plotted against the permuted data, by rank. The intersection of the true data with the upper quartile is shown, which we recommend as a significance cutoff.}
\usage{
fiRankPlot(
  permutedvalues,
  quantiledata,
  xlimitmin = 0,
  xlimitmax = 500,
  ylimitmin = -10,
  ylimitmax = 0,
  labelverticaladjust = 1.05,
  labelhorizontaladjust = 1.05,
  indvPermScoresOn = TRUE,
  logOn = TRUE
)
}
\arguments{
\item{permutedvalues}{a tibble of featureRanks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}

\item{xlimitmin}{a numerical value for the minimum x value in the figure, default is 0}

\item{xlimitmax}{a numerical value for the maximum x value in the figure, default is 500 (this is effectively the total number of features to include)}

\item{ylimitmin}{a numerical value for the minimum y value in the figure, default is -10}

\item{ylimitmax}{a numerical value for the maximum y value in the figure, default is 0, but this will vary largely depending on your dataset, set it larger then narrow down}

\item{labelverticaladjust}{a numerical value specifying the vertical (y-axis) adjustment of the intersection with upper quartile label}

\item{labelhorizontaladjust}{a numerical value specifying the horizontal (x-axis) adjustment of the intersection with upper quartile label}

\item{indvPermScoresOn}{logical TRUE or FALSE, whethere or not to include the lines for each individual permutation feature importance score (shown in light blue)}

\item{logOn}{logical TRUE or FALSE, whether or not to use the log or unlogged data. Default to TRUE.}
}
\value{
A plot of log feature importance score by feature rank showing the upper quartile intersection which we recommend to use for a cutoff.
}
\description{
Generate Feature Importance Score-Rank Plot
This function generates a figure which shows the true observed data plotted against the permuted data, by rank. The intersection of the true data with the upper quartile is shown, which we recommend as a significance cutoff.
}
