% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasets.R
\name{load_permuted_fi}
\alias{load_permuted_fi}
\title{Load feature importance scores for permuted data}
\usage{
load_permuted_fi(
  pathtofile,
  featureImportanceColumnName,
  featureRankColumnName
)
}
\arguments{
\item{pathtofile}{Path to a folder containing minimum 100 csv files (one for each null permutation). Input files must be in .csv format as a table of feature importance scores or Gini Importance scores for each named feature. Must have a minimum of two columns, one named featurename and another named featureImportance. The true non-randomized data .csv file must NOT be in this folder.}

\item{featureImportanceColumnName}{The name of the column in the input .csv file which contains the feature importance scores.}

\item{featureRankColumnName}{The name of the column in the input .csv file which contains the ordered feature ranks. Option: enter "rownames" if the rank is stored as the rownames. Feature ranks should start at 1.}
}
\value{
A tibble containing all of the feature importance score data for the permuted datasets.
}
\description{
This function loads the first random forest results files for the 100 randomized outcome variable null permutations
}
