% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasets.R
\name{load_permuted_fws}
\alias{load_permuted_fws}
\title{Load feature weights for permuted data}
\usage{
load_permuted_fws(pathtofile, featureWeightColumnName)
}
\arguments{
\item{pathtofile}{Path to a folder containing minimum 100 csv files (one for each null permutation). Input files must be in .csv format as a table of feature weights or Gini Importance scores for each named feature. Must have a minimum of two columns, one named featurename and another named featureweight. The true non-randomized data .csv file must NOT be in this folder.}
}
\value{
A tibble containing all of the feature weight data for the permuted datasets.
}
\description{
This function loads the first random forest results files for the 100 randomized outcome variable null permutations
}
