% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasets.R
\name{load_true_fws}
\alias{load_true_fws}
\title{Load feature weight for true data}
\usage{
load_true_fws(pathtofile, featureWeightColumnName)
}
\arguments{
\item{pathtofile}{Path to the input file. Input file must be in .csv format as a table of feature weights or Gini Importance scores for each named feature. Must have a minimum of two columns, one named featurename and another named featureweight.}
}
\value{
A tibble containing all of the feature weight data for the true (observed) dataset.
}
\description{
This function loads the first random forest results file
}
