% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{piECDF}
\alias{piECDF}
\title{Generate a plot of the empirical cumulative density function showing the sum of absolute deviations by fraction of data.}
\usage{
piECDF(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{a tibble of featureRanks, feature names, feature importance scores, log feature importance scores and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A plot of the empirical cumulative density function showing the sum of absolute deviations by fraction of data.
}
\description{
Generate a plot of the empirical cumulative density function showing the sum of absolute deviations by fraction of data.
}
