% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{piHistogram}
\alias{piHistogram}
\title{Generate a histogram showing the sum of absolute deviations by count, the measures used to calculate the pi statistics used in the p-value calculation for the set.
This function generates a histogram illustrating the sum of absolute deviations in the permuted set vs the true (observed) set. It is a visualization of the data used to calculate the p-value for the entire feature set.}
\usage{
piHistogram(permutedvalues, quantiledata)
}
\arguments{
\item{permutedvalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_permuted_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}
}
\value{
A histogram of count by sum of absolute deviations, showing the differences between the null and observed datasets.
}
\description{
Generate a histogram showing the sum of absolute deviations by count, the measures used to calculate the pi statistics used in the p-value calculation for the set.
This function generates a histogram illustrating the sum of absolute deviations in the permuted set vs the true (observed) set. It is a visualization of the data used to calculate the p-value for the entire feature set.
}
