% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{weightRankPlot}
\alias{weightRankPlot}
\title{Generate Weight-Rank Plot
This function generates a figure which shows the true observed data plotted against the permutated data, by rank. The intersection of the true data with the upper quartile is shown, which we recommend as a significance cutoff.}
\usage{
weightRankPlot(
  truevalues,
  permutatedvalues,
  quantiledata,
  xlimitmin = 0,
  xlimitmax = 500,
  ylimitmin = 0,
  ylimitmax = 0.0045
)
}
\arguments{
\item{truevalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_true_fws function}

\item{permutatedvalues}{a tibble of featureRanks, feature names, feature weights, log feature weights and permutation # generated by the load_permutated_fws function}

\item{quantiledata}{a tibble containing: featureRanks, mean, lower, upper, and observed as well as logmean, loglower, logupper, logobserved generated by the calculateQuantiles function}

\item{xlimitmin}{a numerical value for the minimum x value in the figure, default is 0}

\item{xlimitmax}{a numerical value for the maximum x value in the figure, default is 500 (this is effectively the total number of features to include)}

\item{ylimitmin}{a numerical value for the minimum y value in the figure, default is 0}

\item{ylimitmax}{a numerical value for the maximum y value in the figure, default is 0.0045, but this will vary largely depending on your dataset, set it larger then narrow down}
}
\value{
A plot of feature weight by feature rank showing the upper quartile intersection which we recommend to use for a cutoff.
}
\description{
Generate Weight-Rank Plot
This function generates a figure which shows the true observed data plotted against the permutated data, by rank. The intersection of the true data with the upper quartile is shown, which we recommend as a significance cutoff.
}
